requirements.txt
selenium>=4.15.0
keyring>=24.3.0
PyYAML>=6.0  # Optional: for YAML configuration file support